+++ src/link/index.js
@@ -1,3 +1,4 @@
 // @flow
 export { ExternalLink } from "./external-link";
 export { ExternalLinkButton } from "./external-link-button";
+export { LinkButton } from "./link-button";